+++ src/entity.hpp
@@ -7,7 +7,7 @@ public:
     Entity(const std::string& spriteLocation, float xPos, float yPos, sf::Vector2f spriteDims);
 
     sf::Sprite GetSprite() { return sprite_; };
-    sf::Vector2i GetPos() { return sf::Vector2i(xPos_, yPos_); }
+    sf::Vector2f GetPos() { return sf::Vector2f(xPos_, yPos_); }
 
     sf::Vector2f GetSpritePosition() { return sprite_.getPosition(); } // might be unnecessary, because sprite pos should be same as that returned of GetPos()
